move functional test to its own package.
